sunxi: Pine64: add (32-bit) SPL support
When loading U-Boot directly without boot0 or via the USB FEL boot
method, we need to use an SPL to initialize the SoC properly, which
most importantly sets up the DRAM for us.
For the time being we only have a 32-bit binary blob taking care of
this, so let's create a separate 32-bit Pine64 configuration just for
the sake of generating the SPL.
This patch provides a still empty sunxi_dram_init(), which will later
be filled in.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>