sunxi: A64: do an RMR switch if started in AArch32 mode
When we start execution with the 32-bit SPL part, both the FEL code
and the SPL image entry will enter the real U-Boot still in 32-bit
mode.
By using a "magic" instruction, which happens to be an almost-NOP on
AArch64 and a branch on AArch32, we differentiate between being
entered in 64-bit or 32-bit mode.
For 64-bit we proceed with the branch to reset, but in 32-bit mode
we trigger an RMR write to bring the core into AArch64/EL3 and re-enter
U-Boot at CONFIG_SYS_TEXT_BASE.
This allows a 32-bit SPL to eventually start a 64-bit U-Boot.
Adds a menu option to select this alternate BOOT0 boot code.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>