arm: Kconfig: move BOOT0 header choice into sunxi menu
The option to have the U-Boot binary prepended with a header isn't
really a user's choice, but more a requirement for certain boards.
Make the general BOOT0 option invisible and let it be selected by specific
board configurations or menu items.
This introduces the sunxi Allwinner boot0 boot method as a new submenu.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>