video: cfb_console: prepare for fonts with 12 pixels wide characters
Currently the cfb_console only supports fonts with 4 or 8 pixel wide
glyphs. Lifting this restrictions looks hairy, as we would loose the
neat optimization we get from restricting the glyph width to multiples
of 4.
But adding support for 12 pixels is pretty straight-forward, so we do
this to allow bigger fonts for boards usually connected to high
resolution screens.
This also adds a build check to make sure the font matches the known
restrictions.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>