serial: S5P/Samsung: refactor and Kconfig-ize UART selection
Currently the UART used by some Samsung SoCs is selected by the
generic CONFIG_S5P define. This makes it hard to re-use that UART
without pulling in the whole of the S5P code. Also the Exynos SoCs
use this driver, which is in fact a generic Samsung UART driver.

So create a new Kconfig symbol CONFIG_SAMSUNG_UART to reflect this and
change the existing boards to use this symbol.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>