HACK: video_dm: use less brighter white for DM video console
If using a white background with a black font, the resulting contrast
is very high, as the (maximum) white background is very bright and
straining to the eyes.
Let's hack the background color definition to use only 75% of
brightness, so things are much easier to read, especially on bright
screens.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>