ARM: arch-timer: improve FSL erratum workaround
The Allwinner A64 seems to suffer from a similar timer erratum like
the Freescale SoC, which has a software workaround. However when
applying the workaround on the Pine64, I get reproducible hangs in the
SPL DRAM setup code after *resetting* the board (first boot works fine).
It looks like it loops in the workaround code forever.
This is probably due to the CPU running at 24 MHz, the same frequency
as the timer, so the chance of reading the same timer value twice is
rather low.
All we really care about is the timer values not going backwards, so
let's rather check for that instead of insisting on equality.
Also let's copy the algorithm which is used in Linux, which imposes an
upper limit on the number of retries, avoiding an infinite loop.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>