sunxi: disable direct MMC environment for 64 bit boards
Since the dawn of time for the Allwinner support in mainline U-Boot
we store the environment to the SD card and write directly at
544KB from the beginning of the device. This leads to problems when
the U-Boot proper image grows beyond 504KB and eventually overlaps.
With one release of having the environment preferably in a FAT
partition, let's now turn off the MMC variant fallback, so we get back
all the space we need to implement features.
In case someone desperately needs to turn direct MMC environment back
on, let's move that to 896KB (128K below 1MB), so that we won't be
easily limited anymore.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>