sunxi: add generic defconfigs for Allwinner A64 boards
The _defconfig files for the various boards with the Allwinner A64 SoCs
aren't actually that different, as U-Boot itself doesn't really care
about their differences, we can cover them with DTs.
Create two common defconfigs, which carry the DTs of all supported
boards. As the SPL can only be compiled to support one kind of DDR3
(DDR3 or LPDDR3), we can't just use a single unified defconfig for now.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>