sunxi: dts: Pine64: Use fixed regulator for mmc0
Linux kernels before 4.12-rc1 don't know about the AXP803 PMIC, so can't
enable the MMC driver with the current DT anymore, because that now
depends on this regulator.
To restore compatiblity with older kernels (where MMC worked fine
without explicit AXP support), restore the former fixed-regulator.
This is no technical disadvantage, as the MMC power rail is actually
VDD-SYS, which drives most of the rest of the board, so can't be turned
off anyway. To underline this, this DCDC1 rail is marked as
"regulator-always-on" in the DT.

This basically reverts the effect of Linux commit
3f241bfa60bdc9c4fde63fa6664a8ce00fd668c6.
Author: Jagan Teki <jagan@amarulasolutions.com>
Date:   Sat Aug 12 11:10:02 2017 +0530
   arm64: allwinner: a64: pine64: Use dcdc1 regulator for mmc0

Signed-off-by: Andre Przywara <andre.przywara@arm.com>