sunxi: pinctrl: Add A64 DM pinctrl driver
Add a simple pinctrl driver for Allwinner SoCs. This uses the generic
pinctrl bindings.
Instead of holding tables with mappings from the function strings to
the actual per-SoC pinmux bits, this relies on these values to be in
the DT node. We use the already existing and generic "pinmux" property.
This makes this driver actually generic for all Allwinner SoCs, given
that we complement the .dtsi files with those properties.
At the moment this just matches on the A64 compatible string.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>