sunxi: rename and export sunxi_read_socid()
Every Allwinner SoC provides a 16-bit SoC ID, which can be read from a
register in the SRAM controller IP.
At the moment we just use that to print that in the CPU info banner, but
actually we can derive many runtime decision from it.

To simplify later usage, export and rename that function to make it more
obvious what it does.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>