sunxi: gpio_uart_init(): detect SoC at runtime
Every SoC has some default debug UART, that we initialise for allowing
U-Boot to output anything. We initialise the GPIO pins connected to this
UART in a big #ifdef cascade.

Switch this over to detect the SoC at runtime instead, so that the UART
is usable on every SoC, regardless of what it was configured for.
This allows to print out messages in every case.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>