sunxi: dram_sunxi_dw: support both DDR3 and LPDDR3 in one build
At the moment the DRAM type used by a board (DDR3 vs. LPDDR3) is a compile
time decision. At least for the SPL this is actually the only difference
between boards using the same SoC, so detecting this at runtime would
allow less binary build. This is especially useful for FEL booting.

Allow both types of DRAM to be selected in Kconfig. Then start with
DDR3, and retry with LPDDR3 if initialisation fails with the former.

When still only one type of DRAM is selected this should not change the
behaviour.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>