sunxi: dram_sunxi_dw: allow runtime choice for DRAM type
At the moment the DRAM type to be used (LPDDR3/DDR3) is a compile time
decision, controlled by some Kconfig variables. We allow only one of
them to be set and choose the right parameters based on this.

To facilitate runtime detection, allow the DRAM type to be changed at
runtime, by introducing a variable and using its value.

At the moment the value of this variable is still set based on Kconfig
symbols, but this will change later on.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>