sunxi: dram_sunxi_dw: allow runtime choice for DRAM clock frequency value
At the moment the clock frequency used for DRAM chips is a compile time
decision.

Replace uses of the Kconfig symbol in the code with a variable, so this
can later be set at runtime, for instance depending on the SoC or DRAM
type used.

At the moment the value is still initialised by the Kconfig symbol.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>