sunxi: H6: enable 32-bit SPL support
For using the SPL in FEL mode (booting via USB-OTG) we need to run it in
AArch32 mode, as switching to AArch64 (and back) changes to much of the
execution state to easily return into the BootROM.
Split the existing A64 Kconfig stanza into a common one and two separate
symbols for 32 and 64 bit.
Change the existing defconfig to use the new MACH_SUN50I_H6_64 symbol.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>