[WIP] sunxi: H6: Add DDR3-1600 timings
Add a routine to program the timing parameters for DDR3-1600 DRAM chip
connected to the H6 DRAM controller.

The values were gathered from doing back-calculations from a register dump,
trying to match them up with the official JEDEC DDDR3 spec.
If in doubt, the register dump values were taken for now, but the JEDEC
recommendation were added as a comment.

Once the DRAM initialisation works for DDR3 DRAM, we can think about
changing the values to the JEDEC versions, which seem more sane and
align with the DRAM chips datasheets.

== As with the DDR3 core controller driver, this does not work yet. ==

Signed-off-by: Andre Przywara <andre.przywara@arm.com>