sunxi: H6: Use libdram for Eachlink H6 Mini SPL
Those are the parameters read from the image shipped with the
Eachlink H6 Mini TV box. The Android is rooted, so the boot0 image can
be dumped from sector 16 of the eMMC on the Android shell prompt.
This also adds a eachlink_h6_mini_libdram_spl_defconfig to trigger
the SPL build with libdram.