sunxi: H6: disable exception vectors to decrease SPL image size
The SPL for the Allwinner H6 is notoriously close (~500 bytes) to the image
size limit the BootROM imposes on us.
Some needed rework for the DRAM driver will push it over the edge.

Pull the emergency brake and disable the ARMv8 exception vectors to make
some room for now. Ideally there is a better solution.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>