sunxi: introduce RVBAR address into Kconfig
When we switch from AArch32 (which the SoC boots in) to AArch64, we need
to program the entry point into a MMIO register, to magically appear in
the RVBAR register. This address is different between A64/H5 and H6.

Add a Kconfig option to move this differentation out of the code and
pave the way to use this at a second place in a follow-up patch.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>