sunxi: Introduce SUNXI_MULTI_SOC Kconfig symbol
In preparation for using one image for multiple SoCs, introduce a
Kconfig symbol to specify this case.

Guard the sunxi_get_socid() call with this symbol, so that the compiler
can optimise away this call if we only configure for one SoC. That
should keep the code both clean and small if needed.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>