sunxi: Add more SOCID definitions
So far the SOCID_ values are only used for a subset of the supported
parts.

Add the remaining values so that we can start doing autodetection.

This also moves the definition before the inclusion of cpu_sun<x>i.h, so
that we can use them in there as well.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>