sunxi: H6: add generic H6 32-bit SPL defconfig
To be able to use FEL booting (Allwinner's USB-OTG protocol), we need a
32 bit SPL, to easily return into the 32 bit BootROM.
Add a generic defconfig for that purpose, that should work on all boards.
The resulting spl/sunxi-spl.bin can be used with the sunxi-fel tool to
initialise the DRAM via FEL. This allows to load the normal versions for
the rest of the firmware components (ATF, 64 bit U-Boot proper).
This SPL will determine the architecture of the loaded binaries and will
switch the bitness if necessary.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>