spi: sunxi: Support newer SoCs
So far the Allwinner SPI driver just covered the older generation SPI
peripheral found in the A10 and A20 devices.
Newer SoCs use a slightly different device, which is still annoyingly
similar: Too similar to justify a separate driver, but too different
to have a neatly integrated driver.
Embrace the U-Boot spirit and solve this issue with some #ifdefs ;-)

Should the need arise, this can be later turned into runtime decisions,
but so far we will never see the need for a joint driver in a certain
binary build.

The technical changes were based on arch/arm/mach-sunxi/spl_spi_sunxi.c.

Signed-off-by: Oskari Lemmela <oskari@lemmela.net>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>