flush_dcache_range(): align range, add comment
The ARMv7 implementation of flush_dcache_range() checks the range start
and end addresses for being cache line aligned. This is actually not
needed for cleaning the cache, as cleaning does not lose data.
Actually the cache does cleans on its own when evicting cache lines.

To appease the ARMv7 messages, do some rounding in the code, but stick
a comment on top to point out the reason.