invalidate_dcache_range(): add test for buffer coverage
Just invalidating areas of the cache can lose data, when some cache
lines were dirty. This is not a problem when we just prepare some DMA
buffers for the hardware, but we must make sure that we don't
accidentally invalidate other unrelated data.
To guarantee this, the buffer address needs to be cache line aligned -
this is already correctly checked in the invalidate_dcache_range()
implementation.
Also the end of the last cache line being invalidated must still be
within our receive buffer, to make sure we don't invalidate other
data.
Provide a function to wrap the original function and do the check
in there.