sunxi: usb: Add H616 USB support
The H616 has the SIDDQ bit at a different position in the HCI PHY CTL
register.
Let's change our usb_phy_cfg struct to hold a mask of bits to clear
instead of just a bool, so we can configure this appropriately.

Add the H616 compatible string along with this struct on the way.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>