refactor ac_remapping
The original DRAM code from awboot initialises cfg0 with all 0s, in a
separate array. This leads the compiler to put that into BSS, which is
not available before DRAM init in the SPL, as the build system puts its
address into DRAM.
Using memcpy to copy static and read-only data is not clever anyways, so
refactor the whole function to use one "table of tables" (which means it
will end up in .data), and just use a pointer to select the right table.
This also allows to get rid of the memcpy implementation.
Reorder the decision making a bit, to make it clearer what's going on.