move DRAM chip timing setup into separate file
This moves the entire auto_set_timing_para() function into a separate
file, as we do in other mainline DRAM drivers.
This is a bit questionable, as this routine here supports multiple
types, in contrast to mainline, where we focus on one type per file only.
Let's see where this takes us, should be easy to move back.