timings: drop auto_cal_timing(), use ns_to_t()
The auto_cal_timing() function is very similar to what we cal ns_to_t()
in mainline, so let's use that instead, not only because it's shorter.

We lose the ability to use the runtime frequency variable, but this is
not really used anyway, as dram_clk is always set at build time and
never changed.