force use of DRAM chip type from Kconfig
Similar to the frequency value, we do not change the DRAM type at
runtime. Fixing this to a build time constant allows MUCH better
optimisation, without jeopardising the code quality. The compiler will
just remove all code to support the other DRAM types from the binary.

Unfortunately the current (GCC 11) compilers are not able to see the
constant nature of this member, so miss this optimisation opportunity.
Let's help it by making this explicit.