suniv: DRAM: use wait_for_bit_le32()
TODO: blows up code size? Partly reverted in later patches

There is no need for inventing yet another "poll for an MMIO bit set"
function, just use the generic wait_for_bit_le32() implementation.
Since we have multiple users of "set a bit and for it to be cleared",
use a simple wrapper function for that.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>