suniv: DRAM: use speed bin specific DRAM timing setup
A closer look at the DRAM chip timing values reveals that they are
actually very close to the JEDEC defined DDR-400 timings, with a frequency
of a bit less than 200 MHz. However we clock the chips at 156 MHz, which
is the closest to the canonical 166 MHz of the JEDEC DDR-3333 spec.
Combining the longer delay cycles calculated for a higher frequency with
a lower DRAM clock rate leaves quite some performance on the table.

As the DRAM is co-packaged into the same chip, we cannot say which actual
dies got used, but the close proximity probably allows for more ambitious
timings or a faster frequency.

Start easy with calculating the delay cycles based on the actual DRAM
frequency, courtesy of the ns_to_t() function. For frequncies beyond
166 MHz, we use the existing conservative current values.
More optimisations will follow.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>