suniv: DRAM: simplify dram_set_autofresh_cycle()
This function contains quite some obfuscated code to calculate and
program the refresh cycle period. According to the JEDEC DDR DRAM spec,
this is 7.8 us for bigger DRAM arrays, and 15.6 us for smaller ones.
To convert this into clock cycles, this function uses open coded fixed
point arithmetic to avoid division (for which ARMv5 requires specific
runtime library routines). U-Boot provides those routines (probably in
contrast to xboot), so there is no reason to confuse the reader with
some weird division hacks. On top of that this routine seems to support
providing the frequency both in Hz and in MHz, which is just weird.

Clean this all up, by introducing our good old ns_to_t() routine, and
use that to just write what we want: 7.8us or 15.6 us in clock ticks.
Remove the support for feeding in the frequency in Hz, since we are
not using that.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>