sunxi: pmic_bus: support alternative I2C address
Some of the X-Power AXP PMICs can be ordered with an alternative I2C
address, for instance an AXP717 could be shipped with address 0x34 or
with address 0x35.
The datasheets for the AXP707, AXP717 and AXP803 list two possible
addresses, and they are always consecutive. For DM (DT) based drivers
this is no problem, but the Allwinner SPL code relies on a hardcoded
address.

Add a Kconfig variable that will add "1" to the existing address if it
is set.
This enables to use the AXP717 as used on boards with the new Allwinner
A523 chip.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>