sunxi: scan whole DRAM to determine DRAM size
Normally we rely on explicit information from the DRAM initialisation
routine in the SPL to determine the installed DRAM size in U-Boot
proper. As a fallback, if we cannot find the SPL signature, there is a
simple test/scanning routine that does the size detection.

So far this routine just considered the first 2GB of DRAM, because this
was the maximum size U-Boot itself would be interested in. For boards
with more amount of RAM (and not using the SPL), it would put only those
2GB in the device tree, which renders the rest of DRAM unusable.

Scan the maximum supported memory size to catch those cases, and for
instance make all the 4GB visible on newer boards, when the DRAM init is
not done by the SPL.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>