power: pmic: sunxi: separate SPL DCDC3 PMIC rail setup
So far each board that set up the DCDC2 power rail on an AXP PMIC in the
SPL, also set up the DCDC3 rail, so we could guard both with the same
Kconfig check.

To allow changing just one of the two rails, split the guard up, so that
each Kconfig value protects each rail separately.

This enables (potential) support for the AXP323 PMIC, where we must not
touch DCDC2 (since it's dual phased with DCDC1), but want to program
DCDC3.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>