sunxi: A133: DRAM: hardcode MR (mode registers)
Jernej suggested that those parameters should be fixed in the code, and
not be configurable. This is in line with the other DRAM drivers.

Not entirely sure that's a good idea, though, as the MR values seem to
contain some board parameters (CAS latency?), and they provide some good
documentation of what's going on (some of that lost with this patch).