sunxi: A64: enable AXP drivers
So far the 64-bit SoCs were relying on Trusted-Firmware to do the PMIC
setup, which was quite static: enabling most rails, regardless of
whether they are needed or not.

We have a fully DM compliant AXP regulator driver for a while, so that
U-Boot proper can do its own PMIC setup, based on the actual need
(typically for USB, Ethernet, HDMI).

Enable that driver by default for all A64 boards, along with the RSB
driver.

The goal is to allow dropping the ill-fated regulator setup code from
TF-A, so that this can become smaller, and we only enable rails that are
really needed.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>