sunxi: A133: DRAM: switch to writel_relaxed() for MMIO accesses
The writel()/readl() MMIO accessors always carry a barrier instruction,
which is correct from the architectural point of view, and it thus
correctly used this way in Linux.
However for the DRAM code, running on a single in-order core, with the
caches disabled, we mostly don't need the barrier, and even if, then
certainly not on every MMIO access.
Since we do a lot of MMIO accesses in the DRAM code, those extra barrier
instructions pile up, and using writel_relaxed()/readl_relaxed(), which
came without them, save quite some memory for our SRAM constrained code.

Replace writel() with writel_relaxed() and readl() with readl_relaxed()
in some part of the code, to save code size.